filter in myroom works
